% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_strings-random.R
\name{check_unique}
\alias{check_unique}
\title{Remove duplicate strings from character vectors}
\usage{
check_unique(chr_vec, length, digits, upperalpha, loweralpha)
}
\arguments{
\item{chr_vec}{The character vector to check for uniqueness}

\item{length}{The length of the random string}

\item{digits}{Should digits be an option (0-9). TRUE by default.}

\item{upperalpha}{Should capital letters be an option (A-Z). TRUE by default.}

\item{loweralpha}{Should lowercase letters be an option (a-z). TRUE by default.}
}
\value{
The character vector without any duplicates
}
\description{
Remove duplicate strings from character vectors
}
\keyword{internal}
