% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_cli.R
\name{cl_text_input}
\alias{cl_text_input}
\title{Ask Text Questions via the Console}
\usage{
cl_text_input(prompt, .envir = parent.frame(1))
}
\arguments{
\item{prompt}{A character string with the question to be asked. Passed into
\code{\link[cli:cli_text]{cli::cli_text()}} and can use its theming syntax.}

\item{.envir}{Used to ensure that \code{\link[cli:cli_text]{cli::cli_text()}} evaluates the prompt,
yes_message, and no_message in the appropriate environment. Expert use
only.}
}
\value{
A character string containing the user's response.
}
\description{
This function allows you to interact with a user by posing questions that
accept text input. It's inspired by \code{\link[usethis:ui-questions]{usethis::ui_yeah()}} and
\code{\link[shiny:textInput]{shiny::textInput()}}. It wraps \code{\link[base:readline]{base::readline()}} with \code{\link[cli:cli_text]{cli::cli_text()}}
allowing you to format questions using the \code{cli} package.
}
\details{
In general, I would recommend using \code{\link{cl_text_action}} as it
accepts code to execute depending on the user's input, but this function is
perhaps more adaptable.
}
\examples{

favorite_cat_question <- function() {
  answer <- cl_text_input("Name your {.strong favorite} cat.")
  if (tolower(answer) != "tucker") {
    cli::cli_alert_danger("\U1F63E Come meet my cat, Tucker,
    and I think you'll change your mind!")
  } else {
    cli::cli_alert_success("Correct answer! \U1F63B")
  }
}

favorite_dog_question <- function() {
  answer <- cl_text_input("Name your {.strong favorite} dog.")
  if (tolower(answer) != "ella") {
    cli::cli_alert_danger("\U1F44E Come meet my dog, Ella,
    and I think you'll change your mind!")
  } else {
    cli::cli_alert_success("Correct answer!\U1F436")
  }
}

if (interactive()) favorite_cat_question()

if (interactive()) favorite_dog_question()

}
\seealso{
Other command-line-tools: 
\code{\link{cl_text_action}()},
\code{\link{cl_yes_no_action}()},
\code{\link{cl_yes_no_lgl}()}
}
\concept{command-line-tools}
