% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_classes.R
\name{cls_abort}
\alias{cls_abort}
\title{Abort Classes with Informative Message}
\usage{
cls_abort(object, expected_class)
}
\arguments{
\item{object}{The object whose class should be checked.}

\item{expected_class}{A character vector of expected (or allowed) classes.}
}
\value{
NA; triggers an error.
}
\description{
This function is adapted from
\href{https://github.com/r-lib/gargle/blob/083acb186791c48294fd89f8d4fcb32cedc72aab/R/utils-ui.R#L172}{gargle}'s
\code{abort_bad_class} function. It provides an informative error message
indicating the object and the class it must be for the function to work
properly.
}
\examples{

if (interactive()) {

  is_df <- function(x) {
    if (is.data.frame(x)) {
      cls_abort(mtcars, expected_class = c("character", "numeric"))
    }
  }

  is_df(mtcars)

}


}
