% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt.R
\name{create_env_pw}
\alias{create_env_pw}
\title{Create environment password}
\usage{
create_env_pw(service)
}
\arguments{
\item{service}{Identifier of the service whose token will be encrypted.}
}
\value{
NA; used for side effects to create environmental variable.
}
\description{
Following the conventions of gargle's \verb{secret_*()} family of functions, this
function will generate a random password that can be added to your R
enviornment. This will allow you to pass the system variable to external
services such as GitHub Actions, which can be useful for encrypting and
decrypting authentication tokens.
}
\examples{

if (interactive()) {
create_env_pw("testing-jdtools")
}


}
