% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_encrypt.R
\name{decrypt_token}
\alias{decrypt_token}
\title{Decrypt an encrypted token file}
\usage{
decrypt_token(service, path, complete = FALSE)
}
\arguments{
\item{service}{Identifier of the service whose token will be encrypted.}

\item{path}{The location of the encrypted token file.}

\item{complete}{Logical: FALSE by default and will return decrypted raw
bytes. TRUE and will decrypt to character string via \code{\link[base:rawConversion]{base::rawToChar()}}.}
}
\value{
Either the raw bytes of the decrypted file or a character string to
pass into authentication functions such as \code{\link[boxr:box_auth]{boxr::box_auth()}}.
}
\description{
Decrypt an encrypted token file
}
