% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{dir_exists_prompt}
\alias{dir_exists_prompt}
\title{Check a directory exists}
\usage{
dir_exists_prompt(.path)
}
\arguments{
\item{.path}{The path where an action is to be executed.}
}
\value{
LOGICAL; TRUE if the directory already exists  \emph{or} TRUE if the user
indicates they want to create the directory (in which case it will be
created in this function). FALSE if the directory does not exist and the
user indicates they don't want to create it.
}
\description{
This function is a simple utility for checking whether a directory exists. It
is used internally with the \verb{file_new_*} family of functions.
}
\examples{

if (interactive()) {
  fs::dir_create("test-folder")
  if (dir_exists_prompt(here("test-folder/test-file.R"))) {
    fs::file_create("test-folder/test-file.R")
  }
}

}
