% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt.R
\name{encrypt_token}
\alias{encrypt_token}
\title{Encrypt a token file}
\usage{
encrypt_token(service, input, destination)
}
\arguments{
\item{service}{Identifier of the service whose token will be encrypted.}

\item{input}{The token file to encrypt, typically a ".json" file.}

\item{destination}{The output directory you would like to store this file,
e.g. in a Shiny Web App's "www" subdirectory.}
}
\value{
NA; used for side effects to create an encrypted token file.
}
\description{
Following
\url{https://gargle.r-lib.org/articles/articles/managing-tokens-securely.html#encrypt-the-secret-file},
this function writes an encrypted version of the input file to a specified
directory, such as a Shiny Web App's "www" subfolder. The supplied
destination path is suffixed with a "secret" folder.
}
