% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_strings-extract.R
\name{extract_after}
\alias{extract_after}
\title{Extract all characters in a string after a match}
\usage{
extract_after(full_string, after_string, trim_spaces = TRUE)
}
\arguments{
\item{full_string}{The full character string to extract from.}

\item{after_string}{A pattern specifying the string should be extracted after this.}

\item{trim_spaces}{logical T/F, if there are any spaces within the extracted
string, they will be removed. Default is TRUE.}
}
\value{
A string
}
\description{
If a string has more than one match, this will only return the first one.
}
\examples{

extract_after("What does the fox say?", after_string = "What does the")

}
