% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_strings-identifiers.R
\name{extract_between}
\alias{extract_between}
\title{Extract a portion of a string between two patterns}
\usage{
extract_between(full_string, after_string, before_string, trim_spaces = TRUE)
}
\arguments{
\item{full_string}{The full character string to extract from.}

\item{after_string}{A pattern specifying the string should be extracted after this.}

\item{before_string}{A pattern specifying the string should be extracted before this.}

\item{trim_spaces}{logical T/F, if there are any spaces within the extracted
string, they will be removed. Default is TRUE.}
}
\value{
A string
}
\description{
If a string has more than one match, this will only return the first one.
}
\examples{

extract_between("What does the fox say?", after_string = "What does the", before_string = "say")

}
