% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_new-file.R
\name{file_new}
\alias{file_new}
\alias{file_new_r}
\alias{file_new_stan}
\alias{file_new_js}
\alias{file_new_css}
\alias{file_new_html}
\alias{file_new_text}
\alias{file_new_md}
\alias{file_new_rmd}
\title{Create a new file}
\usage{
file_new(path, ext, open = interactive())

file_new_r(path, open = interactive())

file_new_stan(path, open = interactive())

file_new_js(path, open = interactive())

file_new_css(path, open = interactive())

file_new_html(path, open = interactive())

file_new_text(path, open = interactive())

file_new_md(path, open = interactive())

file_new_rmd(path, open = interactive())
}
\arguments{
\item{path}{A character vector of the file path to be created.}

\item{ext}{The file extension, e.g. "R" or "Rmd" for .R and .Rmd files,
respectively.}

\item{open}{Logical: whether to open the file for editing. Default calls
\code{\link[base:interactive]{base::interactive()}}.}
}
\value{
NA; used to create and (optionally) open a new file of the specified
type.
}
\description{
These functions are inspired by \code{\link[usethis:use_r]{usethis::use_r()}} to create .R files in the
R/ directory for packages. Unlike \code{use_r()}, these functions allow you to
pass in any path within your RStudio Project. It will then create and
(optionally) open the file within RStudio.

There is both a generic \code{file_new()} function, which accepts paths and the
file extension, as well as several utility functions that simply require a
path. See usage or examples for more details.
}
\examples{

if (interactive()) {
  # Create a new R file, specifying the extension
  file_new(path = "test-folder/test-file", ext = "R")

  # Create a new R file with the shortcut:
  new_r(path = "test-folder/test-file")

  # Create a new Stan file
  file_new(path = "test-folder/test-file", ext = "stan")
  # Alternatively:
  # new_stan(path = "test-folder/test-file")

  # Create a new JavaScript file
  file_new(path = "test-folder/test-file", ext = "js")
  # Alternatively:
  # new_js(path = "test-folder/test-file")

  # Create a new CSS file
  file_new(path = "test-folder/test-file", ext = "css")
  # Alternatively:
  # new_css(path = "test-folder/test-file")

  # Create a new HTML file
  file_new(path = "test-folder/test-file", ext = "html")
  # Alternatively:
  # new_html(path = "test-folder/test-file")

  # Create a new Text file
  file_new(path = "test-folder/test-file", ext = "txt")
  # Alternatively:
  # new_txt(path = "test-folder/test-file")

  # Create a new Markdown file
  file_new(path = "test-folder/test-file", ext = "md")
  # Alternatively:
  # new_md(path = "test-folder/test-file")

  # Create a new RMarkdown file
  file_new(path = "test-folder/test-file", ext = "Rmd")
  # Alternatively:
  # new_rmd(path = "test-folder/test-file")

}


}
