% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{file_open}
\alias{file_open}
\title{Open a File in RStudio}
\usage{
file_open(path, ...)
}
\arguments{
\item{path}{A character vector of the file to be opened.}

\item{...}{Additional arguments to be passed into
\code{\link[rstudioapi:navigateToFile]{rstudioapi::navigateToFile()}}.}
}
\value{
NA; used for side effects to open a file in RStudio.
}
\description{
This function is simply a wrapper around \code{\link[rstudioapi:navigateToFile]{rstudioapi::navigateToFile()}} and
will open the file specified by the \code{path} argument in the current RStudio
window.
}
