% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_strings-random.R
\name{get_individual_string}
\alias{get_individual_string}
\title{Get individual random string}
\usage{
get_individual_string(
  length = 6,
  digits = TRUE,
  upperalpha = TRUE,
  loweralpha = TRUE
)
}
\arguments{
\item{length}{The length of the random string}

\item{digits}{Should digits be an option (0-9). TRUE by default.}

\item{upperalpha}{Should capital letters be an option (A-Z). TRUE by default.}

\item{loweralpha}{Should lowercase letters be an option (a-z). TRUE by default.}
}
\value{
One random string
}
\description{
Get individual random string
}
\keyword{internal}
