% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_strings-random.R
\name{get_possible_values}
\alias{get_possible_values}
\title{Get possible values for random string generation}
\usage{
get_possible_values(digits = TRUE, upperalpha = TRUE, loweralpha = TRUE)
}
\arguments{
\item{digits}{Should digits be an option (0-9). TRUE by default.}

\item{upperalpha}{Should capital letters be an option (A-Z). TRUE by default.}

\item{loweralpha}{Should lowercase letters be an option (a-z). TRUE by default.}
}
\value{
A character vector of some combination of digits, uppercase, and lowercase letters.
}
\description{
Get possible values for random string generation
}
\keyword{internal}
