% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Value Matching}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{vector or \code{NULL}; the values to be checked.}

\item{table}{vector or \code{NULL}, the vectors to be checked against.}
}
\value{
A logical vector indicating if a match was \strong{not} located for each
element of \code{x}.
}
\description{
This operator is simply the inverse of the native \code{\%in\%} operator, and will
return true if the left-hand side is \strong{not} in the right hand side. See the
examples for more details.
}
\examples{

best_cats <- c("Tucker", "Oliver", "Mr. Loon")

# Returns TRUE, obviously
"Tucker" \%in\% best_cats

# Returns FALSE (because Tucker is the best cat)
"Tucker" \%nin\% best_cats


# Create a vector of vowels
vowels <- c("a", "e", "i", "o", "u")

# Create a sentence
sentence <- "Sushi is great!"
# Split sentence into individual letters
sentence_letters <- strsplit(sentence, "")[[1]]

# Subset the sentence for all vowels
sentence_letters[sentence_letters \%in\% vowels]
# Subset the sentence for all consonants
sentence_letters[sentence_letters \%nin\% vowels]


# Vector of numbers 1 through 10
all_nums <- seq(0,10)
# Vector of a few numbers between 1 through 10
sub_nums <- c(1,3,6)

# Which numbers between 1 and 10 are in the subset
all_nums[all_nums \%in\% sub_nums]

# Which numbers between 1 and 10 are not in the subset
all_nums[all_nums \%nin\% sub_nums]

}
