% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_how-many.R
\name{how_many.data.frame}
\alias{how_many.data.frame}
\title{This is the S3 method for class 'data.frame'}
\usage{
\method{how_many}{data.frame}(x, column)
}
\arguments{
\item{x}{Object of class 'data.frame'}

\item{column}{The column of the data frame to be indexed into.}
}
\value{
An integer indicating how many unique observations are in the object.
}
\description{
This is the S3 method for class 'data.frame'
}
\examples{

how_many(mtcars, "cyl") # Subset dataframe with a character string for column name
how_many(mtcars, cyl) # Subset dataframe without needing to quote the column name

}
