% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_how-many.R
\name{how_many}
\alias{how_many}
\title{How Many Unique Observations}
\usage{
how_many(x, ...)
}
\arguments{
\item{x}{Object such as a data.frame or vector}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
An integer indicating how many unique observations are in the object.
}
\description{
This generic function calculates how many unique observations are in a vector
It is identical to \code{length(unique(x))}. However, this function can be used on
either a vector or a data frame with a specified column. See the examples for
more details.
}
\examples{

how_many(mtcars$cyl) # Pass in vector directly
how_many(mtcars, "cyl") # Subset dataframe with a character string for column name
how_many(mtcars, cyl) # Subset dataframe without needing to quote the column name

}
