% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{overwrite_file}
\alias{overwrite_file}
\title{Check a file exists and, if so, prompt to overwrite}
\usage{
overwrite_file(.path)
}
\arguments{
\item{.path}{The path where an action is to be executed.}
}
\value{
LOGICAL; TRUE if the file does not exist and should be "overwritten"
\emph{or} TRUE if the user indicates they want to overwrite the file. FALSE if
the user indicates they don't want to overwrite the file.
}
\description{
This function is a simple utility for checking whether a file exists and
should be overwritten. It is used internally with the \verb{file_new_*} family of
functions.
}
\examples{

if (interactive()) {
  fs::dir_create("test-folder")
  fs::file_create("test-folder/test-file.R")
  if (overwrite_file("test-folder/test-file.R")) {
    fs::file_create("test-folder/test-file.R")
  }
}

}
