% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_strings-random.R
\name{rand_str}
\alias{rand_str}
\title{Generate Random Strings}
\usage{
rand_str(
  num_strings = 10,
  length = 6,
  digits = TRUE,
  upperalpha = TRUE,
  loweralpha = TRUE,
  unique = TRUE
)
}
\arguments{
\item{num_strings}{Number of strings}

\item{length}{Length of strings}

\item{digits}{Logical: TRUE and digits will be in string, FALSE and they won't.}

\item{upperalpha}{Logical: TRUE and capital letters will be in string, FALSE and they won't.}

\item{loweralpha}{Logical: TRUE and lowercase will be in string, FALSE and they won't.}

\item{unique}{Logical: TRUE and strings will be unique, FALSE and they may not be.}
}
\value{
A character vector of random strings defined by the parameters as above.
}
\description{
Generate random strings of any length using a combination of uppercase or
lowercase letters and digits 0-9.
}
\examples{

user_ids <- data.frame(id = rand_str(num_strings = 50))

}
