% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rstudio.R
\name{rs}
\alias{rs}
\title{Restart the R Session in RStudio}
\usage{
rs(command = "")
}
\arguments{
\item{command}{A command (as a character string) to be run after restarting.}
}
\value{
NA; used for side effects.
}
\description{
This function is simply a wrapper around \code{\link[rstudioapi:restartSession]{rstudioapi::restartSession()}} and,
if called within RStudio, will restart the R session.
}
