% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_save-load.R
\name{set_object_path}
\alias{set_object_path}
\title{Set global path option for saving/loading objects}
\usage{
set_object_path(path)
}
\arguments{
\item{path}{A directory within the current RStudio Project}
}
\value{
NA; used for side effects to set the global option
\code{jdtools.rds.path}.
}
\description{
This function will set the global option for saving/loading objects with
\code{jdtools} functions so you do not have to specify the \code{path} parameter when
calling \code{\link{save_object}} and \code{\link{load_object}}.
}
\examples{

if (interactive()) {

fs::dir_create(here::here("test-folder"))
jdtools::set_object_path("test-folder")

}

}
\seealso{
Other utils-save-load: 
\code{\link{get_object_path}()},
\code{\link{unset_object_path}()}
}
\concept{utils-save-load}
