% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_strings-identifiers.R
\name{time_string}
\alias{time_string}
\title{Get a Character String of Seconds since January, 1970}
\usage{
time_string(length)
}
\arguments{
\item{length}{Number of characters to make the string.}
}
\value{
A character string based on the current time since January 1st, 1970.
}
\description{
This function returns a character string of user-defined length based on the
current number of seconds since January 1st, 1970. It is useful for
prototyping data storage pipelines where the actual file name may be unknown.
}
\examples{

time_string(length = 4)
time_string(length = 6)
# Will throw a warning
time_string(length = 4646)

}
