% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_save-load.R
\name{unset_object_path}
\alias{unset_object_path}
\title{Unset global path option for saving/loading objects}
\usage{
unset_object_path()
}
\value{
NA; used for side effects to unset the global option
\code{jdtools.rds.path}.
}
\description{
Unset global path option for saving/loading objects
}
\examples{

if (interactive()) {

fs::dir_create(here::here("test-folder"))
# Will get message saying no path is set.
jdtools::unset_object_path("test-folder")
jdtools::set_object_path("test-folder")
# Will actually unset the path.
jdtools::unset_object_path("test-folder")

}

}
\seealso{
Other utils-save-load: 
\code{\link{get_object_path}()},
\code{\link{set_object_path}()}
}
\concept{utils-save-load}
